media-source/SourceBuffer-abort-removed.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>SourceBuffer#abort() for already removed buffer from parent media source</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

async_test(function(t) {
    var mediaSource = new MediaSource();
    mediaSource.addEventListener('sourceopen', function(e){
        var sourceBuffer = mediaSource.addSourceBuffer(mime);
        mediaSource.removeSourceBuffer(sourceBuffer);
        t.step(function() {
            assert_throws({name: 'InvalidStateError'},
                          function() {
                              sourceBuffer.abort();
                          },
                          'SourceBuffer#abort() after removing the SourceBuffer object');
        });
        t.done();
    }, false);
    video.src = window.URL.createObjectURL(mediaSource);
}, 'SourceBuffer#abort : if this object has been removed from the sourceBuffers attribute of the parent media source then throw an INVALID_STATE_ERR exception and abort these steps.');

</script>
</body>
</html>
